refactor: 샘플이미지 query 분리
